Separated out NodePanel to CurrentTopicView and original NodePanel
